+++ toys-core/lib/toys/standard_mixins/highline.rb
@@ -21,10 +21,6 @@
 # IN THE SOFTWARE.
 ;
 
-require "toys/utils/gems"
-Toys::Utils::Gems.activate("highline", "~> 2.0")
-require "highline"
-
 module Toys
   module StandardMixins
     ##
@@ -62,6 +58,9 @@ module Toys
       KEY = ::Object.new.freeze
 
       to_initialize do |*args|
+        require "toys/utils/gems"
+        Toys::Utils::Gems.activate("highline", "~> 2.0")
+        require "highline"
         self[KEY] = ::HighLine.new(*args)
         self[KEY].use_color = $stdout.tty?
       end