+++ toys-core/test/mixins/test_terminal.rb
@@ -22,6 +22,7 @@
 ;
 
 require "helper"
+require "toys/utils/terminal"
 require "toys/standard_mixins/terminal"
 
 describe Toys::StandardMixins::Terminal do
@@ -38,7 +39,7 @@ describe Toys::StandardMixins::Terminal do
       tool "foo" do
         include :terminal
         def run
-          exit(terminal.is_a?(::Toys::Terminal) ? 1 : 2)
+          exit(terminal.is_a?(::Toys::Utils::Terminal) ? 1 : 2)
         end
       end
     end